// src/app/travel-management/interfaces/travel-details.ts
import { ITravel } from "./travel";

export interface TravelDetails {
    id: string;
    departAgency: string;
    arrivalAgency: string;
    bus?: string;
    busDriver?: string;
    plannedDepartDatetime: Date;
    effectiveDepartDatetime?: Date;
    arrivalDatetime?: Date;
    travelType: string;
    travelState: string;
    travelItem?: ITravel
}